requirements.txt
convertdate==2.3.2
hijri-converter==2.1.3
holidays==0.11.2
korean-lunar-calendar==0.2.1
PyMeeus==0.5.11
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
termcolor==1.1.0
